chore(PubSub): Move pool to its own package
